fibroai/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.2
tqdm==4.66.5
nltk==3.9.1
spacy==3.7.6
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.2
umap-learn==0.5.6
hdbscan==0.8.38.post1
bertopic==0.16.4
gensim==4.3.3
octis==1.14.0
contextualized-topic-models==2.6.0
